Refactorize MultiSyncModeSelectorTests
